Started building the infrastructure for profit functions so agents make better decisions; need to link this to actual opportunity costs computing and such and finally I can move on to have species heatmaps that are ready.

+++ src/main/java/uk/ac/ox/oxfish/fisher/log/FishingRecord.java
@@ -16,7 +16,7 @@ public class FishingRecord
     /**
      * hours spent fishing.
      */
-    final private double hoursSpentFishing;
+    final private int hoursSpentFishing;
 
     /**
      * the gear used
@@ -46,7 +46,7 @@ public class FishingRecord
 
 
     public FishingRecord(
-            double hoursSpentFishing, Gear gearUsed, SeaTile tileFished, Catch fishCaught,
+            int hoursSpentFishing, Gear gearUsed, SeaTile tileFished, Catch fishCaught,
             Fisher fisher, int step) {
         this.hoursSpentFishing = hoursSpentFishing;
         this.gearUsed = gearUsed;
@@ -56,7 +56,7 @@ public class FishingRecord
         this.step = step;
     }
 
-    public double getHoursSpentFishing() {
+    public int getHoursSpentFishing() {
         return hoursSpentFishing;
     }
